add flash message support to application
Adicionando flash messages para exibir feedback ao usuário em ações como criação, edição e exclusão de registros

+++ src/Controller/FarmController.php
@@ -24,46 +24,56 @@
             return $this->render('farm/index.html.twig', $data);
         }
 
-         #[Route('/fazenda/adicionar', name: 'farm_add')]
-         public function addFarm(Request $request, EntityManagerInterface $em) : Response
-         {
-            $msg = '';
+        #[Route('/fazenda/adicionar', name: 'farm_add')]
+        public function addFarm(Request $request, EntityManagerInterface $em): Response
+        {
             $farm = new Farm;
             $form = $this->createForm(FarmType::class, $farm);
             $form->handleRequest($request);
 
-            if($form->isSubmitted() && $form->isValid()){
-                $em->persist($farm);
-                $em->flush();
-                $msg ='Fazenda adicionada com sucesso!';
+            if ($form->isSubmitted() && $form->isValid()) {
+
+                try {
+                    $em->persist($farm);
+                    $em->flush();
+
+                    $this->addFlash('success', 'Fazenda adicionada com sucesso!');
+                    return $this->redirectToRoute('farm_index');
+
+                } catch (\Doctrine\DBAL\Exception\UniqueConstraintViolationException $e) {
+                    $this->addFlash('danger', 'Erro: este nome de fazenda já está cadastrado.');
+                }
             }
 
-            $data = [
-                            'title' => 'Adicionar Fazenda',
-                            'form'  => $form,
-                            'msg' => $msg,
-            ];
+            return $this->render('farm/form.html.twig', [
+                'title' => 'Adicionar Fazenda',
+                'form'  => $form
+            ]);
+        }
 
-            return $this->render('farm/form.html.twig', $data);
-         }
 
         #[Route('/fazenda/editar/{id}', name: 'farm_edit')]
         public function editFarm($id, Request $request, EntityManagerInterface $em, FarmRepository $farmRepository) : Response
         {
-            $msg = '';
             $farm = $farmRepository->find($id);
             $form = $this->createForm(FarmType::class, $farm);
             $form->handleRequest($request);
 
-            if($form->isSubmitted() && $form->isValid()){
-                $em->flush();
-                $msg = 'Fazenda Atualizado com sucesso!';
+            if ($form->isSubmitted() && $form->isValid()) {
+                try {
+                    $em->flush();
+                    $this->addFlash('success', 'Fazenda atualizada com sucesso!');
+                    return $this->redirectToRoute('farm_index');
+
+                } catch (\Doctrine\DBAL\Exception\UniqueConstraintViolationException $e) {
+                    $this->addFlash('danger', 'Erro: já existe uma fazenda com este nome.');
+                }
             }
+
 
             $data = [
                         'title' => 'Editar Fazenda',
-                        'form'  => $form,
-                        'msg' => $msg,
+                        'form'  => $form
             ];
 
             return $this->render('farm/form.html.twig', $data);
@@ -73,9 +83,22 @@
         public function removeFarm($id, EntityManagerInterface $em, FarmRepository $farmRepository) : Response
         {
             $farm = $farmRepository->find($id);
+
+            if (!$farm) {
+                $this->addFlash('danger', 'Fazenda não encontrada.');
+                return $this->redirectToRoute('farm_index');
+            }
+
+            if ($farm->getCows()->count() > 0) {
+                $this->addFlash('danger', 'Não é possível excluir: a fazenda possui gado associado.');
+                return $this->redirectToRoute('farm_index');
+            }
+
             $em->remove($farm);
             $em->flush();
 
+            $this->addFlash('success', 'Fazenda apagada com sucesso!');
             return $this->redirectToRoute('farm_index');
         }
+
     }
